+++ src/components/Header.tsx
@@ -107,7 +107,7 @@ export default function Header() {
             </Link>
 
             {/* Desktop Navigation */}
-            <div className="absolute left-1/2 hidden -translate-x-1/2 items-center justify-center gap-3 xl:flex 2xl:gap-5">
+            <div className="absolute left-1/2 hidden -translate-x-1/2 items-center justify-center gap-7 xl:flex 2xl:gap-9">
               {navigationItems.map((item) => (
                 <Link
                   key={item.href}